feat: add `--no-recommended-warning` option
