+++ packages/repack/src/commands/common/config/makeCompilerConfig.ts
@@ -15,6 +15,7 @@ import { getEnvOptions } from './getEnvOptions.js';
 import { getRepackConfig } from './getRepackConfig.js';
 import { loadProjectConfig } from './loadProjectConfig.js';
 import { normalizeConfig } from './normalizeConfig.js';
+import validateRequiredPlugins from './validateRequiredPlugins.js';
 
 interface MakeCompilerConfigOptions {
   args: StartArguments | BundleArguments;
@@ -23,6 +24,7 @@ interface MakeCompilerConfigOptions {
   rootDir: string;
   platforms: string[];
   reactNativePath: string;
+  noRecommendedPlugins?: boolean;
 }
 
 export async function makeCompilerConfig<C extends ConfigurationObject>(
@@ -73,29 +75,16 @@ export async function makeCompilerConfig<C extends ConfigurationObject>(
     normalizeConfig(config, options.platforms[index])
   );
 
-  const packageJson = JSON.parse(
-    readFileSync(join(rootDir, 'package.json'), 'utf-8')
-  );
+  if (!options.noRecommendedPlugins) {
+    const activePlugins = new Set(
+      normalizedConfigs
+        .flatMap((c) => ('plugins' in c ? c.plugins : []))
+        .map((p) => p?.constructor.name)
+        .filter((p): p is string => p !== undefined)
+    );
 
-  const dependencies = Object.keys(packageJson.dependencies);
-  const activePlugins = new Set(
-    normalizedConfigs
-      .flatMap((c) => ('plugins' in c ? c.plugins : []))
-      .map((p) => p?.constructor.name)
-  );
-
-  dependencies
-    .filter((d) => DEPENDENCIES_WITH_SEPARATE_PLUGINS[d])
-    .forEach((d) => {
-      const requiredPlugin = DEPENDENCIES_WITH_SEPARATE_PLUGINS[d];
-
-      if (!activePlugins.has(requiredPlugin.plugin)) {
-        console.warn(
-          `${bold('WARNING:')} Detected ${bold(d)} package which requires ${bold(requiredPlugin.plugin)} plugin but it's not configured. ` +
-            `Please add the following to your configuration file. \nRead more https://github.com/callstack/repack/tree/main/packages/${requiredPlugin.path}/README.md.`
-        );
-      }
-    });
+    validateRequiredPlugins(rootDir, activePlugins, options.bundler);
+  }
 
   return normalizedConfigs as C[];
 }
