+++ packages/repack/src/commands/options.ts
@@ -71,6 +71,10 @@ export const startCommandOptions = [
       '[DEPRECATED] Path to a bundler config file, e.g webpack.config.js. Please use --config instead.',
     parse: (val: string) => path.resolve(val),
   },
+  {
+    name: '--no-recommended-plugins',
+    description: 'Disables recommended plugins',
+  },
 ];
 
 export const bundleCommandOptions = [
@@ -161,5 +165,9 @@ export const bundleCommandOptions = [
     description:
       '[DEPRECATED] Path to a bundler config file, e.g webpack.config.js. Please use --config instead.',
     parse: (val: string) => path.resolve(val),
+  },
+  {
+    name: '--no-recommended-plugins',
+    description: 'Disables recommended plugins',
   },
 ];
