+++ packages/repack/src/commands/webpack/start.ts
@@ -54,6 +54,7 @@ export async function start(
     rootDir: cliConfig.root,
     platforms: args.platform ? [args.platform] : detectedPlatforms,
     reactNativePath: cliConfig.reactNativePath,
+    noRecommendedPlugins: args.noRecommendedPlugins,
   });
 
   // expose selected args as environment variables
