+++ website/src/latest/api/cli/bundle.mdx
@@ -128,6 +128,12 @@ Path to a bundler config file, e.g rspack.config.js.
 
 Path to a bundler config file, e.g webpack.config.js.
 
+### `--no-recommended-plugins`
+
+- Type: `boolean`
+
+Disables recommended plugins warnings.
+
 ### `-h`, `--help`
 
 Display help for command.
