+++ website/src/latest/api/cli/start.mdx
@@ -114,6 +114,12 @@ Path to a bundler config file, e.g webpack.config.js.
 
 Path to a bundler config file, e.g webpack.config.js.
 
+### `--no-recommended-plugins`
+
+- Type: `boolean`
+
+Disables recommended plugins warnings.
+
 ### `-h`, `--help`
 
 Display help for command.
